Got LDC form validation working
