+++ exoctk/exoctk_app/__init__.py
@@ -1,2 +1,3 @@
 from . import app_exoctk
 from . import log_exoctk
+from . import form_validation
